pyproject.toml
[project]
name = "making-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.13",
    "fastapi>=0.115.13",
    "groq>=0.29.0",
    "langchain-text-splitters>=0.3.8",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=4.1.0",
    "uvicorn[standard]>=0.34.3",
]
